sign-language/requirements.txt
fastapi>=0.128,<1.0
uvicorn[standard]>=0.40,<1.0
jinja2>=3.1,<4.0
numpy>=2.0,<3.0
pillow>=11.0,<12.0
scikit-learn>=1.7,<2.0
joblib>=1.4,<2.0
kagglehub>=0.3,<1.0
httpx>=0.28,<1.0
tqdm>=4.67,<5.0
